#81: improvements to filesystem util

+++ sxpat/utils/filesystem.py
@@ -1,7 +1,8 @@
+from typing import Iterable, NoReturn
+
 import os
 import shutil
 import tempfile
-from typing import Iterable
 
 
 __all__ = ['FS']
@@ -13,6 +14,8 @@ class FS:
 
         @authors: Marco Biasion
     """
+
+    def __new__(cls) -> NoReturn: raise NotImplementedError(f'{cls.__qualname__} is a utility class and as such cannot be instantiated')
 
     @classmethod
     def exists(cls, path: str) -> bool:
@@ -30,41 +33,46 @@ class FS:
     def rmdir(cls, path: str, recursive: bool = False) -> None:
         """Remove the directory (recursively if wanted). Does nothing if the directory does not exist."""
         path = os.path.normpath(path)
-        if FS.exists(path): (shutil.rmtree if recursive else os.rmdir)(path)
+        if os.path.exists(path): (shutil.rmtree if recursive else os.rmdir)(path)
 
     @classmethod
-    def cleandir(cls, path: str) -> None:
-        """Creates or empties the directory."""
+    def emptydir(cls, path: str) -> None:
+        """Empties an existing directory."""
         path = os.path.normpath(path)
-        FS.mkdir(path)
+
+        if not os.path.exists(path): raise FileNotFoundError(f'{path} does not exists')
+        if not os.path.isdir(path): raise NotADirectoryError(f'{path} is not a directory')
+
         for _path in FS.listdir(path):
             if os.path.isfile(_path) or os.path.islink(_path): os.remove(_path)
             elif os.path.isdir(_path): shutil.rmtree(_path)
 
     @classmethod
     def listdir(cls, path: str) -> Iterable[str]:
-        """Returns a list of paths representing all files in the given folder."""
+        """Returns an iterable of paths corresponding to the contents of the given folder."""
         path = os.path.normpath(path)
         return (os.path.join(path, file) for file in os.listdir(path))
 
-    @classmethod
-    def open(cls, path: str, mode: str):
-        """TODO: should this also create the directory or not?"""
-        raise NotImplementedError()
+    # @classmethod
+    # def open(cls, path: str, mode: str):
+    #     """TODO: should this also create the directory or not?"""
+    #     raise NotImplementedError()
 
-        path = os.path.normpath(path)
-        directory = os.path.dirname(path)
+    #     path = os.path.normpath(path)
+    #     directory = os.path.dirname(path)
 
-        cls.mkdir(directory)
-        return open(path, mode)
+    #     cls.mkdir(directory)
+    #     return open(path, mode)
 
     @classmethod
     def open_tmp(directory: str = None, delete: bool = False, binary: bool = False):
-        """ Create a temporary file on the filesystem. \n
-            If `directory` is given, the file will be created in that directory (created if missing). \n
-            If `delete` is `True` the file will be deleted once it is closed. \n
+        """
+            Create a temporary file on the filesystem.  
+            If `directory` is given, the file will be created in that directory (created if missing).  
+            If `delete` is `True` the file will be deleted once it is closed.  
             If `binary` is `True` the file will be opened in binary mode.
         """
+
         if directory is not None:
             directory = os.path.normpath(directory)
             FS.mkdir(directory)
@@ -75,3 +83,18 @@ class FS:
             dir=directory,
             delete=delete
         )
+
+    @classmethod
+    def copy(src_path: str, dst_path: str, exists_ok: bool = False) -> None:
+        """
+            Copies a file or an entire directory from source to destination.  
+            Raises an exception if `exists_ok` is false and `dst_path` already exists.
+        """
+
+        src_path = os.path.normpath(src_path)
+        dst_path = os.path.normpath(dst_path)
+
+        if not exists_ok and os.path.exists(dst_path): raise FileExistsError(f'{dst_path} already exists')
+
+        if os.path.isdir(src_path): shutil.copytree(src_path, dst_path, dirs_exist_ok=True)
+        else: shutil.copyfile(src_path, dst_path, follow_symlinks=True)
